feat : 리뷰 서비스에 리뷰삭제 권한 확인 기능 추가

+++ src/main/java/sparta/ifour/movietalk/domain/reviews/service/ReviewService.java
@@ -47,17 +47,21 @@ public class ReviewService {
     public void updateReview(ReviewRequestDto requestDto, Long reviewId, User user) { // 리뷰 수정
         Review review = getReviewById(reviewId);
 
-        if(!user.getLoginId().equals(review.getUser().getLoginId()))
-            throw new AccessDeniedException("다른 사용자가 작성한 리뷰는 수정할 수 없습니다.");
-
+        validateAuthor(review, user);
         review.update(requestDto.getTitle(), requestDto.getContent(), requestDto.getRatingScore());
     }
 
     @Transactional
-    public void deleteReview(Long reviewId) { // 리뷰 삭제
+    public void deleteReview(Long reviewId, User user) { // 리뷰 삭제
         Review review = getReviewById(reviewId);
 
+        validateAuthor(review, user);
         reviewRepository.delete(review);
+    }
+
+    public void validateAuthor(Review review, User user) {
+        if(!user.getLoginId().equals(review.getUser().getLoginId()))
+            throw new AccessDeniedException("다른 사용자가 작성한 리뷰는 수정할 수 없습니다.");
     }
 
     @Transactional
